Calm down the possibility of a buffer overflow.
Still the maximum values do not fit exactly into the double,
though way better than the previous overflow while passing a
large qint64 as qint32 into the function. Fixes some progress
issues I've noticed with very large files that I downloaded.

+++ src/libs/kdtools/kdupdaterfiledownloader.cpp
@@ -62,7 +62,7 @@
 using namespace KDUpdater;
 using namespace QInstaller;
 
-static double calcProgress(qint32 done, qint32 total)
+static double calcProgress(qint64 done, qint64 total)
 {
     return total ? (double(done) / double(total)) : 0;
 }
